# Conflicts:
#	src/main/java/be/kuleuven/cs/jli40d/client/Client.java
#	src/main/java/be/kuleuven/cs/jli40d/core/model/Card.java
#	src/main/java/be/kuleuven/cs/jli40d/core/model/Game.java
#	src/main/java/be/kuleuven/cs/jli40d/core/model/GameMove.java

+++ src/main/java/be/kuleuven/cs/jli40d/core/GameHandler.java
@@ -26,7 +26,7 @@ public interface GameHandler extends Remote, Serializable
      * @throws RemoteException
      * @throws GameNotFoundException When the game is not found.
      */
-    boolean isStarted( String token, int gameID ) throws InvalidTokenException, RemoteException, GameNotFoundException;
+    boolean isStarted( String token, long gameID ) throws InvalidTokenException, RemoteException, GameNotFoundException;
 
     /**
      * Returns true if it's the server determines the players (identified
@@ -41,7 +41,7 @@ public interface GameHandler extends Remote, Serializable
      * @throws RemoteException
      * @throws GameNotFoundException When the game is not found.
      */
-    boolean myTurn( String token, int gameID ) throws InvalidTokenException, RemoteException, GameNotFoundException;
+    boolean myTurn( String token, long gameID ) throws InvalidTokenException, RemoteException, GameNotFoundException;
 
 
     /**
@@ -59,7 +59,7 @@ public interface GameHandler extends Remote, Serializable
      * @throws RemoteException
      * @throws GameNotFoundException When the game is not found.
      */
-    GameMove getNextMove( String token, int gameID, int nextGameMoveID ) throws
+    GameMove getNextMove( String token, long gameID, long nextGameMoveID ) throws
             InvalidTokenException,
             RemoteException,
             GameNotFoundException;
@@ -77,7 +77,7 @@ public interface GameHandler extends Remote, Serializable
      * @throws GameNotFoundException    When the game is not found.
      * @throws InvalidGameMoveException When the move is invalid.
      */
-    void sendMove( String token, int gameID, GameMove move ) throws
+    void sendMove( String token, long gameID, GameMove move ) throws
             InvalidTokenException,
             RemoteException,
             GameNotFoundException,
